Move departements list examples from /regions to /departements
