+++ components/departements/examples/departements-list.js
