+++ components/regions/regions.js
@@ -1,17 +1,14 @@
-import FaList from 'react-icons/lib/fa/list'
 import FaTag from 'react-icons/lib/fa/tag'
 
 import Summary from '../summary'
 import TechnicalDoc from '../technical-doc/technical-doc'
 
 import ByName from './examples/by-name'
-import DepartementsList from './examples/departements-list'
 
 import doc from './doc'
 
 const examples = [
-  {title: 'Recherche par nom', id: 'name', icon: <FaTag />},
-  {title: 'Liste de départements', id: 'departements-list', icon: <FaList />}
+  {title: 'Recherche par nom', id: 'name', icon: <FaTag />}
 ]
 
 const Regions = () => (
@@ -21,7 +18,6 @@ const Regions = () => (
     <TechnicalDoc {...doc} />
 
     <ByName />
-    <DepartementsList />
   </div>
 )
